Wipe the data after resizing the userdata partition
If we don't wipe it, old data might be still there.
Or the device will fail to boot up by flashing the device using update.img
when there is partition table change. This is because MASKROM mode update
will not erase the userdata and the userdata sommehow can not be mounted.

Change-Id: I8c16352fcdc4fe0e8953c03ce2adddddd22b724e